+ upwind scheme + immutable flux and big matrices + improved scalar number density function + improved plotting

+++ src/Fluxes/FluxFunctionsCoordinate.jl
@@ -12,7 +12,7 @@
 
         flux = ((-p064 + p164)*(-u0 + u1)*(-(x0^2/2) + x1^2/2)*(-y0 + y1)*(-z0 + z1)*(-phi0 + phi1))
 
-        return Float32(flux)
+        return flux
 
     end
 
@@ -28,7 +28,7 @@
 
         flux = (1/2)*(sqrt(m^2 + p064^2) - sqrt(m^2 + p164^2))*(t0 - t1) * x * (y0 - y1) * (z0 - z1) * (u0 * sqrt(1 - u0^2) - u1 * sqrt(1 - u1^2) - 2*acot_mod(u0) + 2*acot_mod(u1))*(sin(phi0) - sin(phi1))
 
-        return Float32(flux)
+        return flux
 
     end
 
@@ -44,7 +44,7 @@
 
         flux = (1/2)*(-sqrt(m^2 + p064^2) + sqrt(m^2 + p164^2))*(t0 - t1) * (x0 - x1) * (z0 - z1) * (u0 * sqrt(1 - u0^2) - u1 * sqrt(1 - u1^2) - 2*acot_mod(u0) + 2*acot_mod(u1))*(cos(phi0) - cos(phi1))
 
-        return Float32(flux)
+        return flux
 
     end
 
@@ -60,7 +60,7 @@
 
         flux = (1/4)*((sqrt(m^2 + p064^2) - sqrt(m^2 + p164^2))*(t0 - t1)*(u0^2 - u1^2)*(x0^2 - x1^2)*(y0 - y1)*(phi0 - phi1))
 
-        return Float32(flux)
+        return flux
 
     end
